nfcorpus_ndcg/requirements.txt
opensearch-py == 2.3.2
beir == 1.0.0
PyYAML == 6.0